+++ mlperf_common/frameworks/base.py
@@ -34,29 +34,3 @@ class ProfilerHandler:
 
     def pop_nvtx(self):
         raise NotImplementedError
-
-class BaseMPICommunicationHandler(CommunicationHandler):
-    def __init__(self, comm=None, **kwargs):
-        super().__init__(**kwargs)
-        self.comm = comm
-        import numpy as np
-        self.np = np
-
-    def _get_comm(self):
-        if self.comm is None:
-            from mpi4py import MPI
-            self.comm = MPI.COMM_WORLD
-
-        return self.comm
-
-    def barrier(self, sync_group=None):
-        c = self._get_comm() if sync_group is None else sync_group
-        # NOTE: MPI_Barrier is *not* working reliably at scale. Using MPI_Allreduce instead.
-        #c.Barrier()
-        val = self.np.ones(1, dtype=np.int32)
-        result = self.np.zeros(1, dtype=np.int32)
-        c.Allreduce(val, result)
-
-    def global_rank(self):
-        c = self._get_comm()
-        return c.Get_rank()
